attempt to reconnect to stream on error
Summary:
Sometimes (not very often), when I initially load up ci.khanacademy.org
wiht the debugger network panel open, I see that the /deploy/stream request was
cancelled.  It stays cancelled, but if I run the setupStream() function in the
JS console, it seems to work. I don't know what's causing this, and on
inspection of ka-ci:~/deploy-logs/server_error.log, I think we're getting redis
"Socket closed on remote end" stack traces when this happens.... I wasn't able
to repro locally, and also not reliably on ci.khanacademy.org

Until I figure why this is happening, this is a temporary client-side patch to
deal with the event stream dying.

Reviewers: benkomalo, alpert

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D594