Try deploying new versions after previous failures
Summary: Also deploys on startup if the deployed version is out of date.

Test Plan:
If a deploy fails, Mr doesn't try to deploy the same version again but does if she is restarted.
Mr tries to deploy on startup if the contents of staging.version.txt doesn't match the current commit ID.

Reviewers: joel

Reviewed By: joel

Differential Revision: http://phabricator.khanacademy.org/D791